+++ r-diagnostics.tf
@@ -10,6 +10,8 @@ resource "azurerm_virtual_machine_scale_set_extension" "azure_monitor_agent" {
 }
 
 resource "azurerm_monitor_data_collection_rule_association" "dcr" {
+  count = var.azure_monitor_data_collection_rule_enabled ? 1 : 0
+
   name                    = local.dcr_name
   target_resource_id      = azurerm_linux_virtual_machine_scale_set.linux_vmss.id
   data_collection_rule_id = var.azure_monitor_data_collection_rule_id