+++ variables-logs.tf
@@ -1,6 +1,13 @@
+variable "azure_monitor_data_collection_rule_enabled" {
+  description = "`true` to enable Data Collection Rule ID"
+  type        = bool
+  default     = true
+}
+
 variable "azure_monitor_data_collection_rule_id" {
   description = "Data Collection Rule ID from Azure Monitor for metrics and logs collection"
   type        = string
+  default     = ""
 }
 
 variable "azure_monitor_agent_version" {