+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Added
+  * AZ-985: Add option to disable Data Collection Rule
+
 # v7.1.1 - 2023-01-13
 
 Fixed